create constants.h file. small fixes

+++ src/parser.cpp
@@ -6,13 +6,13 @@ Parser parseSerialString(String string) {
   int position = string.indexOf(':');
   String key = string.substring(0, position);
   String value = string.substring(position + 1, string.length());
-  return {key.toInt(), value};
+  return {(unsigned int)key.toInt(), value};
 }
 
 void parseSerialValue(String strValue, String *buf) {
   char terminator = ',';
   unsigned int bufIndex = 0;
-  int position = 0;
+  unsigned int position = 0;
 
   while (position < strValue.length()) {
     int positionNext = strValue.indexOf(terminator, position + 1);
